Fix cast

+++ lm/builder/dump_counts_main.cc
@@ -21,7 +21,7 @@ int main(int argc, char *argv[]) {
   util::ReadCompressed counts(util::OpenReadOrThrow(argv[1]));
   util::scoped_fd vocab_file(util::OpenReadOrThrow(argv[2]));
   lm::builder::VocabReconstitute vocab(vocab_file.get());
-  uint8_t order = boost::lexical_cast<uint8_t>(argv[3]);
+  unsigned int order = boost::lexical_cast<unsigned int>(argv[3]);
   std::vector<char> record(sizeof(uint32_t) * order + sizeof(uint64_t));
   while (std::size_t got = counts.ReadOrEOF(&*record.begin(), record.size())) {
     UTIL_THROW_IF(got != record.size(), util::Exception, "Read " << got << " bytes at the end of file, which is not a complete record of length " << record.size());
